Extract helper functions to a library
Removed everything from the main file and put it into a library
